requirements.txt
confluent-kafka==2.6.1
psycopg2-binary==2.9.10
requests==2.32.3
python-dotenv==1.0.1
